Refactor TransactionDetails layout component structure with composition
- Used composition to create separate BitcoinTransactionDetails and EthereumTransactionDetails components that pass their specific Other Details section information into the generic TransactionDetails component
- Refactored the TransactionDetails screen to use the SingleLineListItem and MultiLineListItem components simplifying the entire layout structure.
